+++ src/components/bluepad32/uni_console.c
@@ -66,6 +66,11 @@ static struct {
     struct arg_end* end;
 } set_bluetooth_enabled_args;
 
+static struct {
+    struct arg_int* idx;
+    struct arg_end* end;
+} disconnect_device_args;
+
 static int list_devices(int argc, char** argv) {
     // FIXME: Should not belong to "bluetooth"
     uni_bluetooth_dump_devices_safe();
@@ -192,7 +197,25 @@ static int del_bluetooth_keys(int argc, char** argv) {
     return 0;
 }
 
+static int disconnect_device(int argc, char** argv) {
+    int idx;
+    int nerrors = arg_parse(argc, argv, (void**)&disconnect_device_args);
+    if (nerrors != 0) {
+        arg_print_errors(stderr, disconnect_device_args.end, argv[0]);
+        return 1;
+    }
+
+    idx = disconnect_device_args.idx->ival[0];
+    if (idx < 0 || idx >= CONFIG_BLUEPAD32_MAX_DEVICES)
+            return 1;
+
+    uni_bluetooth_disconnect_device_safe(idx);
+    return 0;
+}
+
 static void register_bluepad32() {
+    char buf[32];
+
     mouse_set_args.value = arg_dbl1(NULL, NULL, "<value>", "Global mouse scale factor. Higher means faster");
     mouse_set_args.end = arg_end(2);
 
@@ -207,6 +230,11 @@ static void register_bluepad32() {
     set_bluetooth_enabled_args.enabled =
         arg_int1(NULL, NULL, "<0 | 1>", "Whether to enable Bluetooth incoming connections");
     set_bluetooth_enabled_args.end = arg_end(2);
+
+    snprintf(buf, sizeof(buf)-1, "<0-%d>", CONFIG_BLUEPAD32_MAX_DEVICES);
+    disconnect_device_args.idx =
+        arg_int1(NULL, NULL, buf, "Device index to disconnect");
+    disconnect_device_args.end = arg_end(2);
 
     const esp_console_cmd_t cmd_list_devices = {
         .command = "list_devices",
@@ -290,6 +318,13 @@ static void register_bluepad32() {
         .func = &del_bluetooth_keys,
     };
 
+    const esp_console_cmd_t cmd_disconnect_device = {
+        .command = "disconnect",
+        .help = "Disconnects a gamepad/mouse/etc.",
+        .hint = NULL,
+        .func = &disconnect_device,
+    };
+
     ESP_ERROR_CHECK(esp_console_cmd_register(&cmd_set_gap_security_level));
     ESP_ERROR_CHECK(esp_console_cmd_register(&cmd_get_gap_security_level));
     ESP_ERROR_CHECK(esp_console_cmd_register(&cmd_set_gap_periodic_inquiry));
@@ -300,6 +335,7 @@ static void register_bluepad32() {
     ESP_ERROR_CHECK(esp_console_cmd_register(&cmd_list_bluetooth_keys));
     ESP_ERROR_CHECK(esp_console_cmd_register(&cmd_del_bluetooth_keys));
     ESP_ERROR_CHECK(esp_console_cmd_register(&cmd_set_bluetooth_enabled));
+    ESP_ERROR_CHECK(esp_console_cmd_register(&cmd_disconnect_device));
 }
 
 void uni_console_init(void) {